Made simple design UI for homepage, login and registration
hey guys just made a rough UI for the homepage, login, and register. Feel free to build upon it and make any changes. Sylas and AP if you guys have made the homepage or login, etc and its a different design ill adjust the design for your pages. Cheers guys.