Clean Up and Implementation of Study Session Functionality
Got rid of the old homepage controller i know sylas already did that on main but this branch is like 40 commits behind its pretty bad. anyways i made study mode fxml and controller and the buttons on the homepage resemble our final UI more now. we are yet to add the bottom 3 buttons but ill let AP do that. for now im focusing on making everything work so stuff like the log out button being next to the title can be fixed later