Cleaning up UI and Update 2 for Study Mode Functionality
figured out how to seperate using spacing, buttons and text on the top header bar so did that. study mode now asks for the subject you are studying and how long you wish to study for. there is a bug at the moment with pausing the timer so that should be fixed by the next commit