Added Quote Cycling during Study Session
Added a function that gives a new quote every 15 seconds during a study session (we can change this to ai later) and will keep generating them until the session is over.
small bug ive noticed is that if you pause the timer it still keeps counting, so it will keep generating quotes. When i get some more time I will sync the timers or if someone has free time feel free to do so.